Move body generator for internally tagged enums to its own module
Cut-paste
